Renaming classes related to the resolving of script components.
